server-producer/src/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-SocketIO==4.3.0
future==0.18.2
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.1
MarkupSafe==1.1.1
monotonic==1.5
oauthlib==3.1.0
python-engineio==3.12.1
python-socketio==4.5.1
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
